
\name{CNOGpro-internal}
\alias{addState}
\alias{calculateProb}
\alias{copynumberBootstrap}
\alias{getGCinwindows}
\alias{gethits}
\alias{getMembership}
\alias{getreadcounts}
\alias{getsequence}
\alias{HMMcopynumber}
\alias{insertRow}
\alias{isbetween}
\alias{makeViterbimat}
\alias{readGenBank}
\alias{rgammapois}
\alias{sampleChromosome}
\alias{setupEmissionMatrix}
\alias{setupTransitionMatrix}
\alias{solvePR}
\alias{spans}
\alias{updateGeneTable}

\title{Internal CNOGpro Functions}
\description{
 Internal CNOGpro functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
