\name{summaryCNOGpro}
\alias{summaryCNOGpro}
\title{
Summarize the \code{CNOGpro} experiment
}
\description{
Prints basic parameters of the experiment.
}
\usage{
summaryCNOGpro(experiment)
}
\arguments{
  \item{experiment}{
An object of type \code{CNOGpro}
}
}
\author{
Ola Brynildsrud
}
\examples{
data(carsonella)
summaryCNOGpro(carsonella)
}